add reference storage implementation that works without aggregates
