+++ vault/src/database.js
@@ -19,6 +19,7 @@ function getDatabase (name) {
 function createDatabase (name) {
   var db = new Dexie(name)
 
+  // Version 3 adds an `aggregates` table without any further changes.
   db.version(3).stores({
     keys: '++,type',
     events: 'eventId',
